backend/pyproject.toml
[project]
name = "normatik"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    # Web
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.30",

    # ORM / DB
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.13",

    # Auth
    "pyjwt>=2.9",
    "pwdlib[argon2]>=0.2",
    "pyotp>=2.9",
    "qrcode>=7.4",

    # Criptografia de campos sensíveis
    "cryptography>=42",

    # Rate limiting
    "slowapi>=0.1.9",

    # Async tasks
    "celery[redis]>=5.4",
    "redis>=5.0",

    # E-mail
    "aiosmtplib>=3.0",
    "jinja2>=3.1",

    # Storage S3 / MinIO
    "aioboto3>=13.0",

    # PDF (certificados ISO 17025)
    "reportlab>=4.2",

    # Excel (motor GUM)
    "openpyxl>=3.1",

    # Validação
    "pydantic[email]>=2.7",
    "pydantic-settings>=2.3",

    # Utilitários
    "python-multipart>=0.0.9",
    "python-slugify>=8.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[tool.ruff]
target-version = "py312"
line-length = 88
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = ["E501"] # Ignorar linha muito longa (já temos o formatter)

[tool.ruff.lint.isort]
known-first-party = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
